refactor: split out MainContentBody component
